responsiveness for all devices
