+++ code/package.json
@@ -5,6 +5,8 @@
   "dependencies": {
     "@emotion/react": "^11.6.0",
     "@emotion/styled": "^11.6.0",
+    "@material-ui/core": "^4.12.3",
+    "@material-ui/icons": "^4.11.2",
     "@mui/material": "^5.1.1",
     "@reduxjs/toolkit": "^1.6.2",
     "babel-eslint": "^10.0.3",
