+++ code/src/components/AddTask/AddTask.css
@@ -1,11 +1,3 @@
-:root {
-  --primary: #1b1b24;
-  --secondary: #ba68c8;
-  --task-one: #f48db7;
-  --task-two: #864c78;
-  --task-three: #70c4bf;
-}
-
 .form-container {
   border: 2px solid black;
   box-shadow: 6px 6px 1px 1px #000;
@@ -20,8 +12,6 @@
 .input-span {
   display: inline-block;
   position: relative;
-  /* margin-left: auto;
-  margin-right: auto; */
   width: 90%;
 }
