+++ code/src/components/TaskList/TaskList.css
@@ -1,3 +1,12 @@
+.empty-list-section {
+  color: #9e9e9e;
+  padding: 5em;
+}
+
+.emoji-span {
+  margin-left: 3.5em;
+}
+
 .list-section {
   width: 100%;
   margin-left: auto;
